Convert commuter/s3 tests to jest
